Fix input handler leak in Probe and consolidate is_dovi detection
  avpipe.Probe leaked the InputHandler opened by extractCodecInfoForProbe:
  the handle was discarded with _ and never closed, leaving the underlying
  resource (file descriptor or connection) open for the lifetime of the
  process. Fix by storing the handle and deferring Close() so it is released
  after C.probe returns — at which point C.probe has already opened its own
  handle and no longer depends on the pre-extraction handle remaining open.

  Consolidate Dolby Vision detection in is_dovi: dvh1/dvhe codec tags identify
  a stream as Dolby Vision by definition (regardless of whether side data is
  present), so the tag check that lived inline in prepare_video_encoder is
  moved into is_dovi alongside the existing AV_PKT_DATA_DOVI_CONF side-data
  check. The call site simplifies to a single is_dovi(in_stream) call.

  Simplify postCloseFailOpener in tests: replace the maxCycles parameter with
  a simpler "fail after any close" rule (closes > 0) that correctly models the
  content-fabric constraint. Add open/close count assertions to
  TestProbeMVHEVC_VideoLayout to pin the expected handle lifecycle.

  Move probe tests from avpipe_test.go into avpipe_probe_test.go.
